# api/contract/openapi.yaml
openapi: 3.0.0
paths:
...
  /{sessionId}/presence:
    post:
      operationId: MiscController_setPresence
      parameters:
        - name: sessionId
          required: true
          in: path
          schema: {}
      responses:
        '201':
          description: ''
      security:
        - x-api-key: []
      summary: Set presence (composing, recording, available, etc.)
      tags:
        - Misc
  /{sessionId}/presence/subscribe:
    post:
      operationId: MiscController_subscribePresence
      parameters:
        - name: sessionId
          required: true
          in: path
          schema: {}
      responses:
        '201':
          description: ''
      security:
        - x-api-key: []
      summary: Subscribe to presence updates of a contact
      tags:
        - Misc
  /{sessionId}/labels:
    get:
      operationId: MiscController_getLabels
      parameters:
        - name: sessionId
          required: true
          in: path
          schema: {}
      responses:
        '200':
          description: ''
      security:
        - x-api-key: []
      summary: Get labels info (delivered via events)
      tags:
        - Misc
  /{sessionId}/labels/chat/{jid}/{labelId}/add:
    post:
      operationId: MiscController_addChatLabel
      parameters:
        - name: sessionId
          required: true
          in: path
          schema: {}
      responses:
        '201':
          description: ''
      security:
        - x-api-key: []
      summary: Add label to a chat
      tags:
        - Misc
  /{sessionId}/labels/chat/{jid}/{labelId}/remove:
    post:
      operationId: MiscController_removeChatLabel
      parameters:
        - name: sessionId
          required: true
          in: path
          schema: {}
      responses:
        '201':
          description: ''
      security:
        - x-api-key: []
      summary: Remove label from a chat
      tags:
        - Misc
  /{sessionId}/labels/message/{jid}/{messageId}/{labelId}/add:
    post:
      operationId: MiscController_addMessageLabel
      parameters:
        - name: sessionId
          required: true
          in: path
          schema: {}
      responses:
        '201':
          description: ''
      security:
        - x-api-key: []
      summary: Add label to a message
      tags:
        - Misc
  /{sessionId}/labels/message/{jid}/{messageId}/{labelId}/remove:
    post:
      operationId: MiscController_removeMessageLabel
      parameters:
        - name: sessionId
          required: true
          in: path
          schema: {}
      responses:
        '201':
          description: ''
      security:
        - x-api-key: []
      summary: Remove label from a message
      tags:
        - Misc
  /{sessionId}/privacy:
    get:
      operationId: MiscController_getPrivacy
      parameters:
        - name: sessionId
          required: true
          in: path
          schema: {}
      responses:
        '200':
          description: ''
      security:
        - x-api-key: []
      summary: Get all privacy settings
      tags:
        - Misc
  /{sessionId}/privacy/last-seen:
    post:
      operationId: MiscController_updateLastSeen
      parameters:
        - name: sessionId
          required: true
          in: path
          schema: {}
      responses:
        '201':
          description: ''
      security:
        - x-api-key: []
      summary: Update last-seen privacy
      tags:
        - Misc
  /{sessionId}/privacy/online:
    post:
      operationId: MiscController_updateOnline
      parameters:
        - name: sessionId
          required: true
          in: path
          schema: {}
      responses:
        '201':
          description: ''
      security:
        - x-api-key: []
      summary: Update online status privacy
      tags:
        - Misc
  /{sessionId}/privacy/profile-picture:
    post:
      operationId: MiscController_updateProfilePicture
      parameters:
        - name: sessionId
          required: true
          in: path
          schema: {}
      responses:
        '201':
          description: ''
      security:
        - x-api-key: []
      summary: Update profile picture privacy
      tags:
        - Misc
  /{sessionId}/privacy/status:
    post:
      operationId: MiscController_updateStatus
      parameters:
        - name: sessionId
          required: true
          in: path
          schema: {}
      responses:
        '201':
          description: ''
      security:
        - x-api-key: []
      summary: Update status/about privacy
      tags:
        - Misc
  /{sessionId}/privacy/read-receipts:
    post:
      operationId: MiscController_updateReadReceipts
      parameters:
        - name: sessionId
          required: true
          in: path
          schema: {}
      responses:
        '201':
          description: ''
      security:
        - x-api-key: []
      summary: Update read receipts privacy
      tags:
        - Misc
  /{sessionId}/privacy/groups:
    post:
      operationId: MiscController_updateGroups
      parameters:
        - name: sessionId
          required: true
          in: path
          schema: {}
      responses:
        '201':
          description: ''
      security:
        - x-api-key: []
      summary: Update groups add privacy
      tags:
        - Misc
  /{sessionId}/newsletters:
    post:
      operationId: MiscController_createNewsletter
      parameters:
        - name: sessionId
          required: true
          in: path
          schema: {}
      responses:
        '201':
          description: ''
      security:
        - x-api-key: []
      summary: Create a new newsletter/channel
      tags:
        - Misc
  /{sessionId}/newsletters/{newsletterJid}:
    get:
      operationId: MiscController_getNewsletterInfo
      parameters:
        - name: newsletterJid
          required: true
          in: path
          schema: {}
        - name: sessionId
          required: true
          in: path
          schema: {}
      responses:
        '200':
          description: ''
      security:
        - x-api-key: []
      summary: Get newsletter metadata
      tags:
        - Misc
    delete:
      operationId: MiscController_deleteNewsletter
      parameters:
        - name: newsletterJid
          required: true
          in: path
          schema: {}
        - name: sessionId
          required: true
          in: path
          schema: {}
      responses:
        '200':
          description: ''
      security:
        - x-api-key: []
      summary: Delete a newsletter
      tags:
        - Misc
  /{sessionId}/newsletters/{newsletterJid}/subscribers:
    get:
      operationId: MiscController_getSubscribers
      parameters:
        - name: newsletterJid
          required: true
          in: path
          schema: {}
        - name: sessionId
          required: true
          in: path
          schema: {}
      responses:
        '200':
          description: ''
      security:
        - x-api-key: []
      summary: Get newsletter subscriber count
      tags:
        - Misc
  /{sessionId}/newsletters/{newsletterJid}/messages:
    get:
      operationId: MiscController_fetchNewsletterMessages
      parameters:
        - name: after
          required: false
          in: query
          schema: {}
        - name: since
          required: false
          in: query
          schema: {}
        - name: count
          required: false
          in: query
          schema: {}
        - name: newsletterJid
          required: true
          in: path
          schema: {}
        - name: sessionId
          required: true
          in: path
          schema: {}
      responses:
        '200':
          description: ''
      security:
        - x-api-key: []
      summary: Fetch newsletter messages
      tags:
        - Misc
  /{sessionId}/newsletters/follow:
    post:
      operationId: MiscController_followNewsletter
      parameters:
        - name: sessionId
          required: true
          in: path
          schema: {}
      responses:
        '201':
          description: ''
      security:
        - x-api-key: []
      summary: Follow a newsletter/channel
      tags:
        - Misc
  /{sessionId}/newsletters/unfollow:
    post:
      operationId: MiscController_unfollowNewsletter
      parameters:
        - name: sessionId
          required: true
          in: path
          schema: {}
      responses:
        '201':
          description: ''
      security:
        - x-api-key: []
      summary: Unfollow a newsletter/channel
      tags:
        - Misc
  /{sessionId}/newsletters/mute:
    post:
      operationId: MiscController_muteNewsletter
      parameters:
        - name: sessionId
          required: true
          in: path
          schema: {}
      responses:
        '201':
          description: ''
      security:
        - x-api-key: []
      summary: Mute a newsletter/channel
      tags:
        - Misc
  /{sessionId}/newsletters/unmute:
    post:
      operationId: MiscController_unmuteNewsletter
      parameters:
        - name: sessionId
          required: true
          in: path
          schema: {}
      responses:
        '201':
          description: ''
      security:
        - x-api-key: []
      summary: Unmute a newsletter/channel
      tags:
        - Misc
  /{sessionId}/newsletters/send:
    post:
      operationId: MiscController_sendNewsletterMessage
      parameters:
        - name: sessionId
          required: true
          in: path
          schema: {}
      responses:
        '201':
          description: ''
      security:
        - x-api-key: []
      summary: Send a message to a newsletter
      tags:
        - Misc
  /{sessionId}/newsletters/{newsletterJid}/name:
    post:
      operationId: MiscController_updateNewsletterName
      parameters:
        - name: newsletterJid
          required: true
          in: path
          schema: {}
        - name: sessionId
          required: true
          in: path
          schema: {}
      responses:
        '201':
          description: ''
      security:
        - x-api-key: []
      summary: Update newsletter name
      tags:
        - Misc
  /{sessionId}/newsletters/{newsletterJid}/description:
    post:
      operationId: MiscController_updateNewsletterDesc
      parameters:
        - name: newsletterJid
          required: true
          in: path
          schema: {}
        - name: sessionId
          required: true
          in: path
          schema: {}
      responses:
        '201':
          description: ''
      security:
        - x-api-key: []
      summary: Update newsletter description
      tags:
        - Misc
  /{sessionId}/newsletters/{newsletterJid}/react:
    post:
      operationId: MiscController_reactNewsletter
      parameters:
        - name: newsletterJid
          required: true
          in: path
          schema: {}
        - name: sessionId
          required: true
          in: path
          schema: {}
      responses:
        '201':
          description: ''
      security:
        - x-api-key: []
      summary: React to a newsletter message
      tags:
        - Misc
  /{sessionId}/blocklist:
    get:
      operationId: MiscController_getBlocklist
      parameters:
        - name: sessionId
          required: true
          in: path
          schema: {}
      responses:
        '200':
          description: ''
      security:
        - x-api-key: []
      summary: Get blocklist
      tags:
        - Misc
  /{sessionId}/device:
    get:
      operationId: MiscController_getDevice
      parameters:
        - name: sessionId
          required: true
          in: path
          schema: {}
      responses:
        '200':
          description: ''
      security:
        - x-api-key: []
      summary: Get connected device information
      tags:
        - Misc
...
info:
  title: Baileys WhatsApp API
  description: Comprehensive REST API for WhatsApp using Baileys library.
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    x-api-key:
      type: apiKey
      in: header
      name: x-api-key
...